requirements.txt
python-telegram-bot==20.5
fastapi
uvicorn